requirements.txt
django==3.0.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
mysqlclient==1.4.6
Pillow==7.1.1
requests==2.23.0